further cleanup modeTransition.js
