+++ index.html
@@ -114,6 +114,7 @@
     <script src="src/device/touchPropagation.js"></script>
     <script src="src/device/layout.js"></script>
     <script src="src/device/modeTransition.js" type="module"></script>
+    <script src="src/device/PinchGestureRecognizer.js" type="module"></script>
     <script src="src/device/videoRecording.js"></script>
     <script src="src/device/distanceScaling.js"></script>
     <script src="src/device/keyboardEvents.js"></script>
